Assignment#03-TT/river-monitoring-service/requirements.txt
paho-mqtt
flask
pyserial